requirements.txt
channels-redis==4.1.0
daphne==4.0.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-extensions==3.2.3
django-redis==5.3.0
graphene-django==3.1.2
pip-chill==1.0.3
psycopg2-binary==2.9.6
pyopenssl==23.2.0
python-dotenv==1.0.0
requests==2.31.0
service-identity==23.1.0
whitenoise==6.5.0
